+++ src/glyphtune/waveforms/waveform.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 from typing import Any, Callable, Literal, override
 import copy
 import numpy as np
-import glyphtune
+from glyphtune import arrays
 
 
 class Waveform(np.lib.mixins.NDArrayOperatorsMixin):
@@ -23,7 +23,7 @@ class Waveform(np.lib.mixins.NDArrayOperatorsMixin):
     [`waveforms.OperationWaveform`][glyphtune.waveforms.OperationWaveform].
     """
 
-    def sample_arr(self, time_array: glyphtune.FloatArray) -> glyphtune.FloatArray:
+    def sample_arr(self, time_array: arrays.FloatArray) -> arrays.FloatArray:
         """Samples audio data given a time variable array.
 
         Args:
@@ -46,7 +46,7 @@ class Waveform(np.lib.mixins.NDArrayOperatorsMixin):
         duration: float,
         start_offset: float = 0,
         channels: int = 2,
-    ) -> glyphtune.FloatArray:
+    ) -> arrays.FloatArray:
         """Samples audio data given time information in seconds.
 
         Args:
@@ -72,7 +72,7 @@ class Waveform(np.lib.mixins.NDArrayOperatorsMixin):
 
     def sample_samples(
         self, sampling_rate: int, count: int, start_offset: int = 0, channels: int = 2
-    ) -> glyphtune.FloatArray:
+    ) -> arrays.FloatArray:
         """Samples audio data given sample count information.
 
         Args:
@@ -138,7 +138,7 @@ class OperationWaveform(Waveform):
         super().__init__()
         self.__operator = operator
         self.__operands = operands
-        self._operator_kwargs = operator_kwargs
+        self.__operator_kwargs = operator_kwargs
 
     @property
     def operator(self) -> Callable[..., Any]:
@@ -153,10 +153,10 @@ class OperationWaveform(Waveform):
     @property
     def operator_kwargs(self) -> dict[str, Any]:
         """A copy of the keyword arguments this waveform passes to its operator when sampled."""
-        return copy.copy(self._operator_kwargs)
+        return copy.copy(self.__operator_kwargs)
 
     @override
-    def sample_arr(self, time_array: glyphtune.FloatArray) -> glyphtune.FloatArray:
+    def sample_arr(self, time_array: arrays.FloatArray) -> arrays.FloatArray:
         sampled_operands = tuple(
             (
                 operand.sample_arr(time_array)
@@ -165,7 +165,7 @@ class OperationWaveform(Waveform):
             )
             for operand in self.__operands
         )
-        result = self.__operator(*sampled_operands, **self._operator_kwargs)
+        result = self.__operator(*sampled_operands, **self.__operator_kwargs)
         if not isinstance(result, np.ndarray):
             raise TypeError("Operator did not return an array")
         return result
@@ -177,7 +177,7 @@ class OperationWaveform(Waveform):
             and type(self) is type(other)
             and self.__operator == other.operator
             and self.__operands == other.operands
-            and self._operator_kwargs == other.operator_kwargs
+            and self.__operator_kwargs == other.operator_kwargs
         )
 
     @override